refactor(all) : 리팩토링 및 코드 정리
리팩토링 및 코드 정리 :
 - BridgeManagerTest 불필요한 공백 제거
 - BridgeGame 패키지 변경
 - validator 상수 사용 방식 변경
